+++ src/constants/index.ts
@@ -1,1 +1,0 @@
-export { CASE_SPLIT_PATTERN } from './caseSplitPattern';