+++ src/predicate/index.ts
@@ -1,3 +1,4 @@
+export { isEqual } from './isEqual.ts';
 export { isNil } from './isNil.ts';
 export { isNotNil } from './isNotNil.ts';
 export { isNull } from './isNull.ts';