+++ src/string/index.ts
@@ -1,3 +1,4 @@
+export { camelCase } from './camelCase.ts';
 export { snakeCase } from './snakeCase.ts';
 export { kebabCase } from './kebabCase.ts';
 export { lowerCase } from './lowerCase.ts';